+++ embassy-nrf/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - added: System OFF support for the nRF54L series.
 - bugfix: usb: don't re-arm OUT endpoints twice per packet, which could silently drop received packets under load.
+- bugfix: usb: apply the nRF52840 Erratum 199 workaround around USBD EasyDMA transfers.
 
 ## 0.11.0 - 2026-06-16